requirements.txt
# LangGraph Agent Dependencies

langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# OpenTelemetry tracing to New Relic
langsmith[otel]>=0.4.25  # LangSmith with OTEL support for proper span attributes
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=0.43b0
opentelemetry-instrumentation-requests>=0.43b0

